Added Embedding Service Integration Test

+++ src/main/java/com/tensorflow/siamese/services/TfModelServingService.java
@@ -8,7 +8,7 @@ public interface TfModelServingService {
 
     void initializeGraph(String path);
 
-    float[] forward(Tensor Image);
+    double[] forward(Tensor Image);
 
     void closeGraph();
 }
